requirements.txt
openai>=1.55.0
streamlit>=1.35.0
requests>=2.31.0
pandas>=2.0.0
matplotlib>=3.7.0
reportlab>=4.0.0
boto3>=1.34.0
python-dotenv>=1.0.0
pyyaml>=6.0
plotly>=5.18.0
pillow>=10.0.0
